Overflow fixes (#35)
Bench 12117882

+++ src/history.h
@@ -53,7 +53,7 @@ struct QuietHistoryTable {
     };
 
     void update(const Position *pos, const Move move, const int16_t bonus);
-    int16_t getScore(const Position *pos, const Move move) const;
+    int getScore(const Position *pos, const Move move) const;
 };
 
 // Tactical history is a history table for tactical moves
@@ -82,7 +82,7 @@ struct TacticalHistoryTable {
     };
 
     void update(const Position *pos, const Move move, const int16_t bonus);
-    int16_t getScore(const Position *pos, const Move move) const;
+    int getScore(const Position *pos, const Move move) const;
 };
 
 // Continuation history is a history table for move pairs (i.e. a previous move and its continuation)
@@ -112,8 +112,8 @@ struct ContinuationHistoryTable {
 
     void updateSingle(const Position *pos, const SearchStack *ss, const int offset, const Move move, const int16_t bonus);
     void update(const Position *pos, const SearchStack *ss, const Move move, const int16_t bonus);
-    int16_t getScoreSingle(const Position *pos, const SearchStack *ss, const int offset, const Move move) const;
-    int16_t getScore(const Position *pos, const SearchStack *ss, const Move move) const;
+    int getScoreSingle(const Position *pos, const SearchStack *ss, const int offset, const Move move) const;
+    int getScore(const Position *pos, const SearchStack *ss, const Move move) const;
 };
 
 // CorrectionHistoryTable is a history table indexed by [pawn-key-index]. It is used to correct eval
